refactor(profile): move account components into account subfolder
