+++ app/components/settings/account/Page.vue
@@ -1,0 +1,8 @@
+<template>
+  <div>
+    <div class="text-4xl -mb-2 mt-3">Account</div>
+    <Divider />
+    <SettingsAccountAvatar />
+    <SettingsAccountForm />
+  </div>
+</template>
